requirements.txt
fastapi
aiosmtplib
jinja2
python-dotenv
uvicorn
kafka-python
pandas 
openpyxl